refactor: main and parser
