+++ src/parsers/IframelyParser.ts
@@ -1,0 +1,16 @@
+import { Parser } from './parser';
+
+export class IframelyParser extends Parser {
+	constructor() {
+		super();
+		this.api = 'http://iframely.server.crestify.com/iframely?url={{{url}}}';
+	}
+
+	process(data: any): { title: string; image: string; description: string } {
+		const title = data.meta?.title || '';
+		const image = data.links[0]?.href || '';
+		let description: string = data.meta?.description || '';
+		description = description.replace(/\n/g, ' ').replace(/\\/g, '\\\\');
+		return { title, image, description };
+	}
+}
